requirements.txt
slackclient
slackeventsapi
PyYAML
